Separate fixed int types from Char, Int, Size

+++ tests/Kit/Compiler/Passes/test_header.h
@@ -6,7 +6,7 @@ struct Struct1 {
 };
 
 typedef struct {
-  int field1;
+  short field1;
   double field2;
 } Struct2;
 
@@ -26,21 +26,21 @@ typedef enum {
 
 // variables
 
-extern int var1;
+extern short var1;
 char var2 = 1;
 unsigned long var3;
 struct Struct1 struct_var1;
 enum Enum1 enum_var1;
-int *pointer_var1;
-int **pointer_var2;
-int (*pointer_var3)(int arg1);
+short *pointer_var1;
+short **pointer_var2;
+short (*pointer_var3)(short arg1);
 
 // functions
 
 void void_func1();
-int int_func1();
-float func_with_args(int arg1, unsigned long long arg2);
+short int_func1();
+float func_with_args(short arg1, unsigned long long arg2);
 struct Struct1 struct_func(struct Struct2 a);
 float * pointer_func(int *arg1);
-void varargs_func(int a, ...);
+void varargs_func(short a, ...);
 long void_func(void);
